Refactored passes into their interface packages
