+++ plume/src/main/kotlin/io/github/plume/oss/passes/structure/TypeDeclPass.kt
@@ -1,7 +1,8 @@
-package io.github.plume.oss.passes
+package io.github.plume.oss.passes.structure
 
 import io.github.plume.oss.Extractor
 import io.github.plume.oss.drivers.IDriver
+import io.github.plume.oss.passes.IProgramStructurePass
 import io.github.plume.oss.util.SootParserUtil
 import io.github.plume.oss.util.SootToPlumeUtil
 import io.shiftleft.codepropertygraph.generated.EdgeTypes
